# backend/openapi_pretty.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Questionnaire Agent",
        "version": "0.1.0"
    },
    "paths": {
        "/api/v1/projects/": {
            "get": {
                "tags": [
                    "projects"
                ],
                "summary": "Read Projects",
                "description": "Retrieve projects.",
                "operationId": "read_projects_api_v1_projects__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    },
                                    "title": "Response Read Projects Api V1 Projects  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "projects"
                ],
                "summary": "Create Project",
                "description": "Create new project.",
                "operationId": "create_project_api_v1_projects__post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{project_id}": {
            "get": {
                "tags": [
                    "projects"
                ],
                "summary": "Read Project",
                "description": "Get project by ID.",
                "operationId": "read_project_api_v1_projects__project_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Project Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "projects"
                ],
                "summary": "Delete Project",
                "description": "Delete a project.",
                "operationId": "delete_project_api_v1_projects__project_id__delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Project Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{project_id}/generate": {
            "post": {
                "tags": [
                    "projects"
                ],
                "summary": "Generate Answers",
                "description": "Trigger background generation of answers for a project's questionnaire.",
                "operationId": "generate_answers_api_v1_projects__project_id__generate_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Project Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Generate Answers Api V1 Projects  Project Id  Generate Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{project_id}/export": {
            "get": {
                "tags": [
                    "projects"
                ],
                "summary": "Export Report",
                "description": "Export all questions and current answers as a CSV, Excel, Word, or PDF file.",
                "operationId": "export_report_api_v1_projects__project_id__export_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Project Id"
                        }
                    },
                    {
                        "name": "format",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "csv",
                            "title": "Format"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Export Report Api V1 Projects  Project Id  Export Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Job": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "job_type": {
                        "type": "string",
                        "title": "Job Type"
                    },
                    "status": {
                        "$ref": "#/components/schemas/JobStatus"
                    },
                    "progress_percent": {
                        "type": "integer",
                        "title": "Progress Percent"
                    },
                    "error_message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Message"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "job_type",
                    "status",
                    "progress_percent",
                    "error_message"
                ],
                "title": "Job"
            },
            "JobStatus": {
                "type": "string",
                "enum": [
                    "QUEUED",
                    "RUNNING",
                    "COMPLETED",
                    "FAILED"
                ],
                "title": "JobStatus"
            },
            "Project": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "client_name": {
                        "type": "string",
                        "title": "Client Name"
                    },
                    "due_date": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Due Date"
                    },
                    "scope_type": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/ProjectScope"
                            }
                        ],
                        "default": "SINGLE_SOURCE"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ProjectStatus"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    },
                    "questionnaires": {
                        "items": {
                            "$ref": "#/components/schemas/Questionnaire"
                        },
                        "type": "array",
                        "title": "Questionnaires",
                        "default": []
                    },
                    "jobs": {
                        "items": {
                            "$ref": "#/components/schemas/Job"
                        },
                        "type": "array",
                        "title": "Jobs",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "client_name",
                    "id",
                    "status",
                    "created_at",
                    "updated_at"
                ],
                "title": "Project"
            },
            "ProjectCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "client_name": {
                        "type": "string",
                        "title": "Client Name"
                    },
                    "due_date": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Due Date"
                    },
                    "scope_type": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/ProjectScope"
                            }
                        ],
                        "default": "SINGLE_SOURCE"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "client_name"
                ],
                "title": "ProjectCreate"
            },
            "ProjectScope": {
                "type": "string",
                "enum": [
                    "SINGLE_SOURCE",
                    "ALL_DOCS"
                ],
                "title": "ProjectScope"
            },
            "ProjectStatus": {
                "type": "string",
                "enum": [
                    "DRAFT",
                    "INDEXING",
                    "READY",
                    "GENERATING",
                    "REVIEW",
                    "OUTDATED",
                    "COMPLETED",
                    "FAILED"
                ],
                "title": "ProjectStatus"
            },
            "Question": {
                "properties": {
                    "section_header": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Section Header"
                    },
                    "question_number": {
                        "type": "string",
                        "title": "Question Number"
                    },
                    "question_text": {
                        "type": "string",
                        "title": "Question Text"
                    },
                    "order_index": {
                        "type": "integer",
                        "title": "Order Index"
                    },
                    "assigned_to": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Assigned To"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "questionnaire_id": {
                        "type": "integer",
                        "title": "Questionnaire Id"
                    },
                    "project_id": {
                        "type": "integer",
                        "title": "Project Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "question_number",
                    "question_text",
                    "order_index",
                    "id",
                    "questionnaire_id",
                    "project_id",
                    "created_at"
                ],
                "title": "Question"
            },
            "Questionnaire": {
                "properties": {
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "project_id": {
                        "type": "integer",
                        "title": "Project Id"
                    },
                    "parsing_status": {
                        "$ref": "#/components/schemas/QuestionnaireStatus"
                    },
                    "version": {
                        "type": "integer",
                        "title": "Version"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "questions": {
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        },
                        "type": "array",
                        "title": "Questions",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "filename",
                    "id",
                    "project_id",
                    "parsing_status",
                    "version",
                    "created_at"
                ],
                "title": "Questionnaire"
            },
            "QuestionnaireStatus": {
                "type": "string",
                "enum": [
                    "PENDING",
                    "IN_PROGRESS",
                    "COMPLETED",
                    "FAILED"
                ],
                "title": "QuestionnaireStatus"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login/access-token"
                    }
                }
            }
        }
    }
}